requirements.txt
# ===============================
# Core Python tools
# ===============================
numpy==1.26.4
scipy==1.16.3
pandas==2.3.3
matplotlib==3.10.7
opencv-python==4.12.0.88
pillow==11.3.0

# ===============================
# Deep learning & ONNX
# ===============================
torch==2.5.1+cu121
torchvision==0.20.1+cu121
onnx==1.19.1
onnxruntime==1.23.2

# ===============================
# YOLO / Object detection
# ===============================
ultralytics==8.3.228
ultralytics-thop==2.0.18
roboflow==1.2.11

# ===============================
# Environment & config
# ===============================
python-dotenv==1.2.1

# ===============================
# Utilities & helpers
# ===============================
requests==2.32.5
PyYAML==6.0.3
tqdm==4.67.1
coloredlogs==15.0.1
filetype==1.2.0
psutil==7.1.3

# ===============================
# Optional visualization / numerical tools
# ===============================
sympy==1.13.1
networkx==3.5
Jinja2==3.1.6
packaging==25.0
protobuf==4.25.8

# ===============================
# Additional required dependencies (detected in environment)
# ===============================
fonttools==4.60.1
requests-toolbelt==1.0.0
python-dateutil==2.9.0.post0
six==1.17.0
contourpy==1.3.3
kiwisolver==1.4.9
pyparsing==3.2.5
pytz==2025.2
charset-normalizer==3.4.4
urllib3==2.5.0
certifi==2025.11.12
colorama==0.4.6
attrs==25.4.0
fsspec==2025.9.0
setuptools==70.2.0
pip==24.2
